Send large message
This commits added a new function `send_all()`, which dealt with
the situation where a message is larger than a socket can `send()` at a
time.

However, this commit did not fix the issue when message is larger than
the buffer. So when the message is too large, only part of the message
(up to the buffer size) will be sent.